Update issue templates (#8482)
Our issue templates required users checking several "preflight checklist" checkboxes. Requiring these fields was an attempt to encourage users to follow these instructions before filing issues. However, they add unecessary clutter to the issues that are formed, and a lot of users either don't find what they're looking for or ignore the information anyway. Convering the information to Markdown instead of as inputs is more convenient for users and for us.